requirements.txt
langchain-community 
langchain-core 
langchain-groq 
langgraph 
pypdf 
chromadb
unstructured
python-docx
langchain-experimental
db-sqlite3
pandas
numpy
sentence-transformers
tabulate
streamlit 